fix: template instances table badge spacing, chevron icon and navigation
